moves defaults into SofAP so that HttpSoftAP can use them too.  Fixes #24.
