+++ lib/SoftAP.js
@@ -20,9 +20,12 @@ function is(cb) {
 		return true;
 	}
 	throw new Error('Invalid callback function provided.');
-};
+}
 
-function SoftAP(opts) {
+function SoftAP(options) {
+	var opts = SoftAP.defaultOptions();
+	SoftAP.assign(opts, options);
+
 	this.keepAlive = opts.keepAlive;
 	this.noDelay = opts.noDelay;
 	this.timeout = opts.timeout;
@@ -32,6 +35,26 @@ function SoftAP(opts) {
 	this.__publicKey = undefined;
 
 	return this;
+}
+
+SoftAP.defaultOptions = function() {
+	var opts = {
+		host: '192.168.0.1',
+		keepAlive: true,
+		timeout: 8000,
+		noDelay: true,
+		channel: 6
+	};
+
+	return opts;
+};
+
+SoftAP.assign = function(opts, options) {
+	if (options && typeof options == 'object') {
+		Object.keys(options).forEach(function _loadOpts(key) {
+			opts[key] = options[key];
+		});
+	}
 };
 
 function checkResponse(err, dat, cb) {
